+++ src/test/java/net/mtrop/doom/test/TestSuite.java
@@ -1,0 +1,13 @@
+package net.mtrop.doom.test;
+
+import net.mtrop.doom.PK3Test;
+import net.mtrop.doom.WadTest;
+import net.mtrop.doom.test.TestUtils.TestDependsOn;
+
+@TestDependsOn({
+	WadTest.class,
+	PK3Test.class,
+})
+public final class TestSuite
+{
+}